services/patient-manager/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ScanHub-UI Patient Manager"
version = "0.0.1"
description = ""
authors = [
    "David Schote <david.schote@brain-link.de>",
]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11.7"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
asyncpg = "^0.27.0"
psycopg2-binary = "^2.9.6"
async-sqlalchemy = "^1.0.0"

[tool.poetry.group.test.dependencies]
coverage = "*"
codecov = "*"
pytest = "*"

[tool.poetry.group.docs.dependencies]
sphinx = "*"
sphinx_rtd_theme = "*"

[tool.poetry.group.lint.dependencies]
ruff = "^0.12.3"
mypy = "^1.16.1"

[tool.setuptools.package-data]
"app" = ["py.typed"]

[tool.ruff]
src = ["app"]
line-length = 120
indent-width = 4
target-version = "py310"

[tool.ruff.lint]
# Enable linters
# W = pycodestyle, F = Pyflakes, S = flake8-bandit, I = isort, N = pep8-naming, D = pydocstyle
select = ["W", "F", "S", "I", "N", "D"]

[tool.ruff.lint.flake8-type-checking]
strict = true

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.mypy]
plugins = ["sqlalchemy.ext.mypy.plugin"]
warn_return_any = false
check_untyped_defs = true
warn_no_return = true
warn_unreachable = true
python_version = "3.10"
ignore_missing_imports = true